# Back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Seguros TM",
        "description": "API para gestão de seguros, clientes, apólices e sinistros",
        "contact": {
            "email": "suporte@segurostm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/cliente/pagamentos": {
            "get": {
                "tags": [
                    "Cliente - Pagamentos"
                ],
                "summary": "Listar todos os pagamentos do cliente",
                "description": "Retorna histórico completo de pagamentos do cliente autenticado",
                "operationId": "ad5b1887062c891c0a0e691f21e32463",
                "responses": {
                    "200": {
                        "description": "Lista de pagamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_pagamento": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "numero_pagamento": {
                                                        "type": "string",
                                                        "example": "PAG-2024-001"
                                                    },
                                                    "valor_parcela": {
                                                        "type": "number",
                                                        "example": 150
                                                    },
                                                    "data_vencimento": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-02-15"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pendente"
                                                    },
                                                    "apolice": {
                                                        "type": "object"
                                                    },
                                                    "metodo_pagamento": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cliente/pagamentos/{pagamento}/registrar": {
            "post": {
                "tags": [
                    "Cliente - Pagamentos"
                ],
                "summary": "Registrar pagamento",
                "description": "Registra o pagamento de uma parcela com upload de comprovante",
                "operationId": "d6b8761c3dac0cf88941e5d4114b717a",
                "parameters": [
                    {
                        "name": "pagamento",
                        "in": "path",
                        "description": "ID do pagamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "metodo_pagamento_id"
                                ],
                                "properties": {
                                    "metodo_pagamento_id": {
                                        "description": "ID do método de pagamento",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "referencia_pagamento": {
                                        "description": "Referência/código da transação",
                                        "type": "string",
                                        "example": "REF-123456"
                                    },
                                    "comprovante": {
                                        "description": "Arquivo do comprovante de pagamento (imagem ou PDF)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pagamento registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pagamento registado com sucesso"
                                        },
                                        "pagamento": {
                                            "properties": {
                                                "id_pagamento": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pago"
                                                },
                                                "data_pagamento": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "valor_pago": {
                                                    "type": "number",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Pagamento já processado ou erro ao registrar"
                    },
                    "403": {
                        "description": "Não autorizado - pagamento não pertence ao cliente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT. Use o token retornado no endpoint de login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}